Abstract the logic for fetching currently trending words
